fix: move user type into lib/types folder
